feat: add dedicated risk management page
